// web/src/generated/api/models/GroupResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cuentas Claras API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Server-owned group identity and settlement policy.
 * @export
 * @interface GroupResponse
 */
export interface GroupResponse {
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    ownerAccountId: string;
    /**
     * 
     * @type {string}
     * @memberof GroupResponse
     */
    settlementPolicy: GroupResponseSettlementPolicyEnum;
}


/**
 * @export
 */
export const GroupResponseSettlementPolicyEnum = {
    OwnerOnly: 'owner_only',
    AnyMember: 'any_member'
} as const;
export type GroupResponseSettlementPolicyEnum = typeof GroupResponseSettlementPolicyEnum[keyof typeof GroupResponseSettlementPolicyEnum];


/**
 * Check if a given object implements the GroupResponse interface.
 */
export function instanceOfGroupResponse(value: object): value is GroupResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('ownerAccountId' in value) || value['ownerAccountId'] === undefined) return false;
    if (!('settlementPolicy' in value) || value['settlementPolicy'] === undefined) return false;
    return true;
}

export function GroupResponseFromJSON(json: any): GroupResponse {
    return GroupResponseFromJSONTyped(json, false);
}

export function GroupResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GroupResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'ownerAccountId': json['owner_account_id'],
        'settlementPolicy': json['settlementPolicy'],
    };
}

export function GroupResponseToJSON(json: any): GroupResponse {
    return GroupResponseToJSONTyped(json, false);
}

export function GroupResponseToJSONTyped(value?: GroupResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'owner_account_id': value['ownerAccountId'],
        'settlementPolicy': value['settlementPolicy'],
    };
}